Fix cxf-2.1 tests after jetty update from 12.0 to 12.1
